+++ .gitignore
@@ -217,7 +217,7 @@ out.log
 /archive
 /subject_armless_for_cmc.osim
 /muscle_outputs.sto
-/motion_tracking_walking_inverse_outputs.sto
+/motion_prescribed_walking_inverse_outputs.sto
 !/squat_to_stand_visualization/squat_to_stand_visualization.png
 /squat_to_stand_visualization/squat_to_stand.mp4
 /squat_to_stand_visualization/squat_to_stand.webm
@@ -225,30 +225,30 @@ out.log
 !/figures/Moco*Diagram.png
 *.batch
 *.batch.sh
-!/results/motion_tracking_walking_cmc_results/README.md
+!/results/motion_prescribed_walking_cmc_results/README.md
 
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_Actuation_force.sto
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_Actuation_power.sto
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_Actuation_speed.sto
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_controls.sto
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_controls.xml
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_Kinematics_dudt.sto
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_Kinematics_q.sto
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_Kinematics_u.sto
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_pErr.sto
-/motion_tracking_walking_cmc_results/motion_tracking_walking_cmc_states.sto
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_Actuation_force.sto
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_Actuation_power.sto
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_Actuation_speed.sto
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_controls.sto
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_controls.xml
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_Kinematics_dudt.sto
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_Kinematics_q.sto
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_Kinematics_u.sto
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_pErr.sto
+/motion_prescribed_walking_cmc_results/motion_prescribed_walking_cmc_states.sto
 /desiredKinematics_padded.sto
 /desiredKinematics_splinefit_accelerations.sto
 /resources/sitToStand_3dof9musc_dgf.osim
 /resources/Rajagopal2016/ground_reaction_before_and_after_processing.svg
 /resources/Rajagopal2016/grf_walk.mot
 /resources/Rajagopal2016/subject_walk_armless_for_cmc.osim
-/motion_tracking_walking_cmc_tasks.xml
+/motion_prescribed_walking_cmc_tasks.xml
 
-!/figures/motion_tracking_walking_inverse_model.png
+!/figures/motion_prescribed_walking_inverse_model.png
 /resources/squat_to_stand_4dof9musc_dgf.osim
 /MocoPaper.dvi
-/code/motion_tracking_walking_cmc_tasks.xml
+/code/motion_prescribed_walking_cmc_tasks.xml
 /paper/MocoPaper.aux
 /paper/MocoPaper.bbl
 /paper/MocoPaper.blg
