+++ figures/motion_prescribed_walking_inverse_model.png
