--- results/motion_tracking_walking_cmc_results/README.md
@@ -1,1 +1,0 @@
-This directory holds results from CMC.
